// src/types/metrics.ts
export type IMetric = {
  unitsPerEm: number;
  ascender: number;
  capHeight: number;
  xHeight: number;
  baseLine: number;
  descender: number;
  hheaAscender: number;
  hheaDescender: number;
  typoAscender: number;
  typoDescender: number;
  winAscender: number;
  winDescender: number;
  xMax: number;
  xMin: number;
  yMax: number;
  yMin: number;
  useTypoMetrics: boolean;
};